Preparacion de vistas - view
