+++ cinemapedia/lib/config/router/app_router.dart
@@ -4,6 +4,7 @@
 /// nuestro proyecto a la web
 
 import 'package:cinemapedia/presentation/screens/screens.dart';
+import 'package:cinemapedia/presentation/views/views.dart';
 import 'package:go_router/go_router.dart';
 
 final appRouter = GoRouter(
@@ -13,7 +14,7 @@ final appRouter = GoRouter(
     GoRoute(
       path: '/',
       name: HomeScreen.name,
-      builder: (context, state) => const HomeScreen(),
+      builder: (context, state) => const HomeScreen( childView: FavoritesView() ),
 
       // esto lo hacemos para que funcione bien en el navegador web
       routes: [
