+++ cinemapedia/lib/presentation/views/views.dart
@@ -1,0 +1,3 @@
+
+export 'home_views/favorites_view.dart';
+export 'home_views/home_view.dart';
